requirements.txt
fastapi
uvicorn
qiskit
qiskit-aer
qiskit-ibm-runtime
pydantic
# qiskit_quantum_knownledge  # Package not found - commenting out
python-dotenv
